# ProjectOCR/static/swagger.yaml
openapi: 3.0.0
info:
  title: Sample Flask API
  version: 1.0.0
  description: API documentation for the Flask application
paths:
  /users:
    get:
      summary: Retrieve all users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}:
    get:
      summary: Retrieve a single user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /DictEntries:
    get:
      summary: Retrieve all dictionary entries
      responses:
        '200':
          description: A list of dictionary entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DictEntry'
    post:
      summary: Create a dictionary entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - category
                - coef
              properties:
                category:
                  type: string
                coef:
                  type: number
      responses:
        '201':
          description: Dictionary entry created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DictEntry'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /upload:
    post:
      summary: Upload a PDF file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - user_id
              properties:
                file:
                  type: string
                  format: binary
                user_id:
                  type: integer
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  file_id:
                    type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /download/{file_id}:
    get:
      summary: Download a PDF file
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: File downloaded successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string
        pdfs:
          type: array
          items:
            $ref: '#/components/schemas/PDFFile'
    DictEntry:
      type: object
      properties:
        category:
          type: string
        coef:
          type: number
    PDFFile:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        user_id:
          type: integer
        categories:
          type: array
          items:
            type: string
        amounts:
          type: array
          items:
            type: number
        emissions:
          type: array
          items:
            type: number
    Error:
      type: object
      properties:
        error:
          type: string